Seperated Resolvers for to different files
