+++ back-end/package.json
@@ -25,12 +25,12 @@
         "ts-node": "^9.0.0"
     },
     "devDependencies": {
-        "@types/express": "^4.17.7",
-        "@types/mongodb": "^3.5.27",
-        "@types/mongoose": "^5.7.36",
-        "@types/node": "^14.6.2",
+        "@types/express": "^4.17.11",
+        "@types/mongodb": "^3.6.3",
+        "@types/mongoose": "^5.10.3",
+        "@types/node": "^14.14.21",
         "graphql-import-node": "0.0.4",
         "tslint": "^6.1.3",
-        "typescript": "^4.0.2"
+        "typescript": "^4.1.3"
     }
 }
